requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
greenlet>=3.1.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ChatGPT Integration
openai>=1.3.0

# Spreadsheet Export
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
openpyxl>=3.1.2

# Data Processing & Validation
pydantic[email]>=2.5.1
pydantic-settings>=2.1.0
pandas>=2.1.3

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
pytest-cov>=4.1.0

# Development Tools
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0

# Logging & Monitoring
python-json-logger>=2.0.7
prometheus-client>=0.18.0 